requirements.txt
# Core libraries
streamlit==1.28.1
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pandas==2.0.3

# Additional utilities
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2

# For mathematical calculations
scipy==1.11.3

# For deployment and web components
streamlit-webrtc==0.47.1
av==10.0.0

# Optional: For better performance
numba==0.58.1